# src/api.yaml
openapi: 3.0.3
info:
  title: HD1 (Holodeck One) API
  description: |
    Revolutionary API-first 3D coordinate system with universal world boundaries.
    Single source of truth for sessions, worlds, objects, and real-time control.
    
    **CORE PRINCIPLES:**
    - Everything API-driven (zero shell commands)
    - Universal coordinate system [-12, +12] on all axes
    - Objects honor world boundaries absolutely
    - Named object persistence with full lifecycle
    - Real-time WebSocket + REST hybrid
    
    **ARCHITECTURE:**
    - Specification drives code generation
    - Build fails if handlers missing
    - Zero manual routing - auto-generated from spec
    
  version: 1.0.0
  contact:
    name: Holodeck One 3D Engine
servers:
  - url: http://localhost:8080/api
    description: Development server
paths:
...
  /sessions/{sessionId}:
    get:
      operationId: getSession
      tags: [Sessions]
      summary: Get session details
      description: Retrieve detailed session information including all objects
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/sessions/get.go"
      x-function: "GetSession"
      
    delete:
      operationId: deleteSession
      tags: [Sessions]
      summary: Terminate session
      description: Permanently delete session and all associated objects
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session terminated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/sessions/delete.go"
      x-function: "DeleteSession"

  # ========================================
  # SESSION GRAPH SESSION GRAPH & ROOM MANAGEMENT CHANNEL MANAGEMENT
  # TCP-level reliability with channel-based architecture
  # ========================================
...
  /sessions/{sessionId}/entities:
    get:
      operationId: listEntities
      tags: [Entities]
      summary: List all entities in session
      description: |
        Returns all PlayCanvas entities within the specified session.
        Entities are the fundamental building blocks of the game engine.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
        - name: tag
          in: query
          required: false
          schema:
            type: string
          description: Filter entities by tag
          example: "player"
        - name: enabled
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by enabled state
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Maximum number of entities to return
      responses:
        '200':
          description: Entities list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  entities:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntitySummary'
                  total:
                    type: integer
                    example: 1
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/entities/list.go"
      x-function: "ListEntities"

    post:
      operationId: createEntity
      tags: [Entities]
      summary: Create new PlayCanvas entity
      description: |
        Creates a new PlayCanvas entity in the specified session.
        Entities are containers for components in the game engine.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 100
                  description: Human-readable entity name
                  example: "PlayerCharacter"
                tags:
                  type: array
                  items:
                    type: string
                  maxItems: 10
                  description: Entity tags for categorization
                  example: ["player", "dynamic"]
                enabled:
                  type: boolean
                  default: true
                  description: Whether entity is active
                position:
                  $ref: '#/components/schemas/Coordinate3D'
                rotation:
                  $ref: '#/components/schemas/Rotation'
                scale:
                  $ref: '#/components/schemas/Scale3D'
      responses:
        '201':
          description: Entity created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  entity_id:
                    type: string
                    example: "entity-abc123"
                  name:
                    type: string
                    example: "PlayerCharacter"
                  playcanvas_guid:
                    type: string
                    description: PlayCanvas internal GUID
                    example: "pc-guid-xyz789"
                  created_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/entities/create.go"
      x-function: "CreateEntity"
  /sessions/{sessionId}/entities/{entityId}:
    get:
      operationId: getEntity
      tags: [Entities]
      summary: Get entity details
      description: |
        Retrieves detailed information about a specific PlayCanvas entity,
        including all components, transform data, and hierarchy relationships.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
        - name: entityId
          in: path
          required: true
          schema:
            type: string
          example: "entity-abc123"
      responses:
        '200':
          description: Entity details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityDetail'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/entities/get.go"
      x-function: "GetEntity"

    put:
      operationId: updateEntity
      tags: [Entities]
      summary: Update entity properties
      description: |
        Updates entity properties including name, tags, enabled state,
        and transform (position, rotation, scale).
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
        - name: entityId
          in: path
          required: true
          schema:
            type: string
          example: "entity-abc123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                  example: "UpdatedPlayerCharacter"
                tags:
                  type: array
                  items:
                    type: string
                  maxItems: 10
                  example: ["player", "static"]
                enabled:
                  type: boolean
                  example: false
                position:
                  $ref: '#/components/schemas/Coordinate3D'
                rotation:
                  $ref: '#/components/schemas/Rotation'
                scale:
                  $ref: '#/components/schemas/Scale3D'
      responses:
        '200':
          description: Entity updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  entity_id:
                    type: string
                    example: "entity-abc123"
                  updated_at:
                    type: string
                    format: date-time
                  changes:
                    type: array
                    items:
                      type: string
                    example: ["name", "enabled", "position"]
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/entities/update.go"
      x-function: "UpdateEntity"

    delete:
      operationId: deleteEntity
      tags: [Entities]
      summary: Delete entity
      description: |
        Permanently removes a PlayCanvas entity from the session.
        Optionally cascades deletion to child entities.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
        - name: entityId
          in: path
          required: true
          schema:
            type: string
          example: "entity-abc123"
        - name: cascade
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Delete child entities recursively
      responses:
        '204':
          description: Entity deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/entities/delete.go"
      x-function: "DeleteEntity"

  # ========================================
  # COMPONENT MANAGEMENT (PLAYCANVAS v3.0)
  # ========================================
...
components:
  schemas:
...
    SessionSummary:
      type: object
      required: [id, created, world_initialized, object_count, status]
      properties:
        id:
          type: string
        created:
          type: string
          format: date-time
        world_initialized:
          type: boolean
        object_count:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [active, inactive, error]
    SessionDetail:
      allOf:
        - $ref: '#/components/schemas/SessionSummary'
        - type: object
          properties:
            objects:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/ObjectDetail'

    # Environment Schemas
...
    EntitySummary:
      type: object
      required: [entity_id, name, enabled, transform]
      properties:
        entity_id:
          type: string
          description: HD1 entity identifier
          example: "entity-abc123"
        name:
          type: string
          description: Human-readable entity name
          example: "PlayerCharacter"
        enabled:
          type: boolean
          description: Whether entity is active
          example: true
        tags:
          type: array
          items:
            type: string
          description: Entity tags for categorization
          example: ["player", "dynamic"]
        transform:
          type: object
          properties:
            position:
              $ref: '#/components/schemas/Coordinate3D'
            rotation:
              $ref: '#/components/schemas/Rotation'
            scale:
              $ref: '#/components/schemas/Scale3D'
        components:
          type: array
          items:
            type: string
          description: Attached component types
          example: ["model", "rigidbody", "script"]
        playcanvas_guid:
          type: string
          description: PlayCanvas internal GUID
          example: "pc-guid-xyz789"
    EntityDetail:
      allOf:
        - $ref: '#/components/schemas/EntitySummary'
        - type: object
          properties:
            component_details:
              type: object
              description: Detailed component configurations
              example:
                model:
                  asset: "model-asset-id"
                  type: "asset"
                  castShadows: true
                rigidbody:
                  type: "dynamic"
                  mass: 1.0
            children:
              type: array
              items:
                type: string
              description: Child entity IDs
              example: ["entity-child-1", "entity-child-2"]
            parent:
              type: string
              description: Parent entity ID
              example: "entity-parent-1"
            created_at:
              type: string
              format: date-time
            last_updated:
              type: string
              format: date-time
    Coordinate3D:
      type: array
      items:
        type: number
        minimum: -12
        maximum: 12
      minItems: 3
      maxItems: 3
      example: [0, 0, 0]
      description: "3D coordinates within world bounds [-12, +12]"
    Scale3D:
      type: array
      items:
        type: number
        minimum: 0.1
        maximum: 10
      minItems: 3
      maxItems: 3
      example: [1, 1, 1]
      default: [1, 1, 1]

    # Component Schemas (PlayCanvas v3.0)
...
    Rotation:
      type: object
      required: [x, y, z, w]
      properties:
        x:
          type: number
          description: "Quaternion X component"
          example: 0.0
        y:
          type: number
          description: "Quaternion Y component"  
          example: 0.0
        z:
          type: number
          description: "Quaternion Z component"
          example: 0.0
        w:
          type: number
          description: "Quaternion W component"
          example: 1.0
      description: "Quaternion rotation representation"

    # ========================================
    # ENTITY LIFECYCLE SCHEMAS (PLAYCANVAS v3.0)
    # ========================================
...
    Success:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...